Added support for Junit 5 and made it accessable to :app module
