--- core/src/test/java/org/kiwix/kiwixmobile/core/utils/MainDispatcherRule.kt
@@ -1,42 +1,0 @@
-/*
- * Kiwix Android
- * Copyright (c) 2026 Kiwix <android.kiwix.org>
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-package org.kiwix.kiwixmobile.core.utils
-
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.test.StandardTestDispatcher
-import kotlinx.coroutines.test.TestDispatcher
-import kotlinx.coroutines.test.resetMain
-import kotlinx.coroutines.test.setMain
-import org.junit.rules.TestWatcher
-import org.junit.runner.Description
-
-@OptIn(ExperimentalCoroutinesApi::class)
-class MainDispatcherRule(
-  private val dispatcher: TestDispatcher = StandardTestDispatcher()
-) :
-  TestWatcher() {
-  override fun starting(description: Description?) {
-    Dispatchers.setMain(dispatcher)
-  }
-
-  override fun finished(description: Description?) {
-    Dispatchers.resetMain()
-  }
-}
